Renamed the package from pusle_wrapper to pulse_controller
Added an optimized profile that shrinks the bin produced from 5.5mb to 695k. This is much closer to the size of pa-mixer (~500k) which this progam aims to be do the same job as.
