+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rupamix"
-version = "0.2.6"
+version = "0.3.0"
 edition = "2021"
 authors = ["Paulo Guimaraes"]
 description = "Rust PA Mixer is a cli tool for interfacing with Pulse Audio"
@@ -8,6 +8,13 @@ license-file = "LICENSE"
 repository = "https://github.com/fitpaulo/rupamix"
 keywords = ["pulse", "audio", "pulseaudio", "cli", "mixer"]
 categories = ["command-line-utilities", "os::linux-apis", "os::unix-apis"]
+
+[profile.optimized]
+inherits = "release"
+codegen-units = 1
+strip = true
+lto = "fat"
+opt-level = "z"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
