+++ src/pulse_controller/server_info_wrapper.rs
